requirements.txt
requests
beautifulsoup4
PyYAML
ollama
flask
google-generativeai